Move Serialise into Git.Objects
because that's what it's really mostly about :-)
